requirements.txt
allure-pytest==2.14.3
pytest==8.4.0
requests==2.32.4
selenium==4.33.0
pytest-selenium==4.1.0
pytest-xdist==3.7.0
jsonschema==4.21.1
webdriver-manager==4.0.1
selene==2.0.0rc9
